requirements.txt
aiokafka==0.8.0
confluent-kafka==2.0.2
dataclass_wizard==0.22.2
faust_streaming==0.10.11
pytest==7.3.1
pyyaml==6.0
prometheus_client==0.16.0
requests==2.28.2
mock==5.0.2